add all functionality as a module
